override log level properly #244
The default behavior of forge is to log debug messages to a debug.log. Torchmaster can be quiet verbose when it comes to logging on debug. Since these logs are usually useless during normal play, they will be force-disabled now using Configurator.setLevel(logger, level). If debug logging is required, launch the game with `-Dtorchmaster.enableDebugLogging=1`